Improve notification settings UX
There is a rare case in which the new "Enable Notifications" button
doesn't trigger a notification permission request.
This is because in firefox there is a special "Blocked Temporarily"
state which is not exposed though the Web API.
Chrome has done a similar thing see issue #12.
Browser developers, btw. Standardise the web, btw.
idiots